Initial Upload
Sample program "clothes.clp" simply tests CLIPS logic. "pressure.clp" is the primary program with all logic outlined in the pdf. Rare bug causes program to sometimes not have output, waiting for grader to finish looking over program and then I hope to revisit and iron out that last bug. Samples of the program working are also included in the pdf, and the program generally runs without a hitch.